// src/lib/types/wms-api/createRateDto.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * ElasticWMS API
 * Warehouse Management System API - Web, RF, and Integration endpoints
 * OpenAPI spec version: 1.0.0
 */
import type { CreateRateDtoCalculationBasis } from './createRateDtoCalculationBasis.js';
import type { CreateRateDtoRateType } from './createRateDtoRateType.js';

export interface CreateRateDto {
  calculationBasis: CreateRateDtoCalculationBasis;
  currency?: string;
  defaultRate: number;
  effectiveDate?: string;
  expiryDate?: string;
  facilityId: string;
  isActive?: boolean;
  maxCharge?: number;
  minCharge?: number;
  rateCode: string;
  rateName: string;
  rateType: CreateRateDtoRateType;
}