// src/lib/types/wms-api/createBillingCycleDto.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * ElasticWMS API
 * Warehouse Management System API - Web, RF, and Integration endpoints
 * OpenAPI spec version: 1.0.0
 */
import type { CreateBillingCycleDtoFrequency } from './createBillingCycleDtoFrequency.js';

export interface CreateBillingCycleDto {
  billingDay: number;
  cycleCode: string;
  cycleName: string;
  facilityId: string;
  frequency: CreateBillingCycleDtoFrequency;
}