// src/lib/types/wms-api/updateInvoiceStatusDto.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * ElasticWMS API
 * Warehouse Management System API - Web, RF, and Integration endpoints
 * OpenAPI spec version: 1.0.0
 */
import type { UpdateInvoiceStatusDtoStatus } from './updateInvoiceStatusDtoStatus.js';

export interface UpdateInvoiceStatusDto {
  notes?: string;
  paidAt?: string;
  status: UpdateInvoiceStatusDtoStatus;
}